+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Generate Yaml template for HA in up-to-date format
 - Add cover as a specific type
 - Add dimmer lights
+- Add climate template
 
 ## 1.17.1 - 2024-10-28
 - Fix formatting float values